Fix uint cast for attribute unique_id
- Cast customId to int when assigning to UniqueId

cube_pc now reads attributes correctly. Still need to fix:
1. Tagged symbols with variable bit lengths (numSymbols=0)
2. pc_kd_color cumulative probability issue

+++ csharp/Draco.Decoder/DracoDecoder.cs
@@ -215,7 +215,7 @@ public class DracoDecoder
             attribute.AttributeType = (GeometryAttributeType)attributeType;
             attribute.DataType = (DataType)dataType;
             attribute.NumComponents = numComponents;
-            attribute.UniqueId = customId;
+            attribute.UniqueId = (int)customId;
 
             attribute.Init(attribute.AttributeType, attribute.DataType, 
                           attribute.NumComponents, pointCloud.NumPoints);
